requirements.txt
# Core Dependencies
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
openai==1.12.0
python-dotenv==1.0.0

# Microsoft Fabric & Power BI
semantic-link-sempy==0.5.0
powerbiclient==3.1.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
pyarrow==15.0.0
deltalake==0.15.0

# API Integrations
requests==2.31.0
httpx==0.26.0
shopify-python-api==12.3.0
stripe==8.0.0
fedex==3.0.0

# Monitoring & Logging
azure-monitor-opentelemetry==1.2.0
opencensus-ext-azure==1.1.13
structlog==24.1.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
tenacity==8.2.3
jinja2==3.1.3

# Development
pytest==8.0.0
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
